requirements.txt
flask==2.2.3
streamlit==1.21.0
scikit-learn==1.2.2
albumentations==1.3.0
segmentation-models-pytorch==0.3.2
opencv-python==4.7.0.72
numpy==1.24.3
pillow==9.5.0
Werkzeug==2.2.2
gdown